fixed bug 'std::invalid_argument: stoi: no conversion' when sat solver returns terminating '0' in a new line

+++ constraints.h
@@ -7,12 +7,12 @@
 /*
  * Generates at AtLeastOne constraints.
  */
-list<map<unsigned int, bool>> at_least_one_constraints(list<list<unsigned int>> indices) {
-    list<map<unsigned int, bool>> constraints;
+list<map<size_t, bool>> at_least_one_constraints(list<list<size_t>> indices) {
+    list<map<size_t, bool>> constraints;
 
     while (!indices.empty()) {
-        list<unsigned int> elems = indices.back();
-        map<unsigned int, bool> at_least_one;
+        list<size_t> elems = indices.back();
+        map<size_t, bool> at_least_one;
 
         while (!elems.empty()) {
             at_least_one.insert({ elems.back(), true });
@@ -29,15 +29,15 @@ list<map<unsigned int, bool>> at_least_one_constraints(list<list<unsigned int>> 
 /*
  * Generates at AtMostOne constraints.
  */
-list<map<unsigned int, bool>> at_most_one_constraints(list<list<unsigned int>> indices) {
-    list<map<unsigned int, bool>> constraints;
+list<map<size_t, bool>> at_most_one_constraints(list<list<size_t>> indices) {
+    list<map<size_t, bool>> constraints;
 
     while (!indices.empty()) {
-        vector<unsigned int> elems(indices.back().size());
+        vector<size_t> elems(indices.back().size());
         copy(indices.back().begin(), indices.back().end(), elems.begin());
 
-        for (unsigned int i = 0; i < elems.size(); i++) {
-            for (unsigned int j = i + 1; j < elems.size(); j++) {
+        for (size_t i = 0; i < elems.size(); i++) {
+            for (size_t j = i + 1; j < elems.size(); j++) {
                 constraints.push_back({{ elems.at(i), false }, { elems.at(j), false }});
             }
         }
